+++ spec/ruby_ami/stream_spec.rb
@@ -26,6 +26,7 @@ module RubyAMI
       mock_target.expects(:receive_data).send(*(times ? [:times, times] : [:at_least, 1])).with &block
       s = ServerMock.new '127.0.0.1', server_port, mock_target
       @stream = Stream.new '127.0.0.1', server_port, lambda { |m| client.message_received m }
+      @stream.run!
       fake_client.call if fake_client.respond_to? :call
       s.join
       @stream.join